What is this?
This is a gorgeous mod which displays everyone's tier from the Vanilla Pvp Tierlist in game. Sadly, the tiers do not update automatically and I have to do it manually because I cannot get the tiers straight from the spreadsheet. *crying sounds*

This mod will display the player's tier like this: "Ht1 | Ooh_Netiyiy," or "Lt5 | Ooh_Netiyiy."

That is all there is to it; smol but very useful!